Module Sapling.Hash

include Tezos_sapling.Core_sig.T_encoding_compare with type t = Tezos_sapling__Core.Client.Hash.t

include Tezos_sapling.Core_sig.T_encoding with type t = Tezos_sapling__Core.Client.Hash.t

type t

val encoding : t Data_encoding.t

val compare : t -> t -> int

include Tezos_sapling.Core_sig.T_bytes with type t := t

val to_bytes : t -> bytes

val of_bytes_exn : bytes -> t

val uncommitted : height:int -> t

val merkle_hash : height:int -> t -> t -> t

Hash function to compute the merkle tree (Pedersen's hash on JubJub curve). Height is the height we are hashing at in the merkle tree. *

val of_commitment : Tezos_sapling__Core.Client.Commitment.t -> t

Hashes and commitments are the same object but are given different types to avoid confusing nodes and leaves. *

val to_commitment : t -> Tezos_sapling__Core.Client.Commitment.t