Module Tezos_raw_protocol_011_PtHangz2

module Alpha_context : sig ... end

module Alpha_services : sig ... end

module Amendment : sig ... end

Only delegates with at least one roll take part in the amendment procedure. It works as follows:

module Apply : sig ... end

This module supports advancing the ledger state by applying operations.

module Apply_results : sig ... end

module Baking : sig ... end

module Blinded_public_key_hash : sig ... end

module Block_header_repr : sig ... end

module Bootstrap_storage : sig ... end

module Cache_costs : sig ... end

Costs function for cache accesses.

module Cache_memory_helpers : sig ... end

module Commitment_repr : sig ... end

module Commitment_storage : sig ... end

module Constants_repr : sig ... end

module Constants_services : sig ... end

module Constants_storage : sig ... end

module Contract_hash : sig ... end

module Contract_repr : sig ... end

module Contract_services : sig ... end

module Contract_storage : sig ... end

module Cycle_repr : sig ... end

module Delegate_services : sig ... end

module Delegate_storage : sig ... end

module Fees_storage : sig ... end

module Fitness_repr : sig ... end

module Fitness_storage : sig ... end

module Fixed_point_repr : sig ... end

module Gas_limit_repr : sig ... end

module Global_constants_costs : sig ... end

Costs function for the global table of constants.

module Global_constants_storage : sig ... end

This module represents access to a global table of constant Micheline values. Users may register a Micheline value in the table, paying the cost of storage. Once stored, contracts source code may reference this value by its hash.

module Init_storage : sig ... end

Functions to setup storage. Used by Alpha_context.prepare.

module Lazy_storage_diff : sig ... end

module Lazy_storage_kind : sig ... end

module Level_repr : sig ... end

module Level_storage : sig ... end

module Liquidity_baking_cpmm : sig ... end

module Liquidity_baking_lqt : sig ... end

module Liquidity_baking_migration : sig ... end

module Liquidity_baking_repr : sig ... end

module Main : sig ... end

module Manager_repr : sig ... end

module Michelson_v1_gas : sig ... end

module Michelson_v1_primitives : sig ... end

module Migration_repr : sig ... end

module Misc : sig ... end

module Nonce_hash : sig ... end

module Nonce_storage : sig ... end

module Operation_repr : sig ... end

module Parameters_repr : sig ... end

module Path_encoding : sig ... end

module Period_repr : sig ... end

module Raw_context : sig ... end

module Raw_context_intf : sig ... end

All context manipulation functions. This signature is included as-is for direct context accesses, and used in Storage_functors to provide restricted views to the context.

module Raw_level_repr : sig ... end

module Receipt_repr : sig ... end

module Roll_repr : sig ... end

module Roll_storage : sig ... end

Basic roll manipulation.

module Sapling_repr : sig ... end

module Sapling_services : sig ... end

module Sapling_storage : sig ... end

module Sapling_validator : sig ... end

module Saturation_repr : sig ... end

module Script_cache : sig ... end

This module manages the cache for smart contracts.

module Script_comparable : sig ... end

module Script_expr_hash : sig ... end

module Script_int_repr : sig ... end

module Script_interpreter : sig ... end

This is the Michelson interpreter.

module Script_interpreter_defs : sig ... end

module Script_ir_annot : sig ... end

Default annotations

module Script_ir_translator : sig ... end

module Script_list : sig ... end

module Script_map : sig ... end

module Script_repr : sig ... end

module Script_set : sig ... end

module Script_string_repr : sig ... end

Strings of printable characters

module Script_tc_errors : sig ... end

module Script_tc_errors_registration : sig ... end

module Script_timestamp_repr : sig ... end

Defines the internal Michelson representation for timestamps and basic operations that can be performed on it.

module Script_typed_ir : sig ... end

module Script_typed_ir_size : sig ... end

This module provides overapproximation of memory footprint for Michelson-related values.

module Script_typed_ir_size_costs : sig ... end

module Seed_repr : sig ... end

module Seed_storage : sig ... end

module Services_registration : sig ... end

module State_hash : sig ... end

module Storage : sig ... end

module Storage_costs : sig ... end

module Storage_description : sig ... end

module Storage_functors : sig ... end

Tezos Protocol Implementation - Typed storage builders.

module Storage_sigs : sig ... end

module Tez_repr : sig ... end

module Time_repr : sig ... end

module Vote_repr : sig ... end

module Vote_storage : sig ... end

module Voting_period_repr : sig ... end

module Voting_period_storage : sig ... end

module Voting_services : sig ... end