Module Alpha_context.Big_map

module Id : sig ... end

val fresh : temporary:bool -> context -> (context * Id.t) Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

val mem : context -> Id.t -> Script_expr_hash.t -> (context * bool) Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

val get_opt : context -> Id.t -> Script_expr_hash.t -> (context * Script.expr option) Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

val exists : context -> Id.t -> (context * (Script.expr * Script.expr) option) Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

val list_values : ?offset:int -> ?length:int -> context -> Id.t -> (context * Script.expr list) Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

list_values ?offset ?length ctxt id lists all values stored in big map id.

The first offset values are ignored (if passed). Negative offsets are treated as 0.

There will be no more than length values in the result list (if passed). Negative values are treated as 0.

The returned context takes into account gas consumption of loading values.

type update = {

key : Script_repr.expr;

key_hash : Script_expr_hash.t;

value : Script_repr.expr option;

}

type updates = update list

type alloc = {

key_type : Script_repr.expr;

value_type : Script_repr.expr;

}