Module Tezos_raw_protocol_011_PtHangz2.Contract_storage

type Tezos_protocol_environment_011_PtHangz2.Error_monad.error +=

| Balance_too_low of Contract_repr.contract * Tez_repr.t * Tez_repr.t

| Counter_in_the_past of Contract_repr.contract * Tezos_protocol_environment_011_PtHangz2.Z.t * Tezos_protocol_environment_011_PtHangz2.Z.t

| Counter_in_the_future of Contract_repr.contract * Tezos_protocol_environment_011_PtHangz2.Z.t * Tezos_protocol_environment_011_PtHangz2.Z.t

| Unspendable_contract of Contract_repr.contract

| Non_existing_contract of Contract_repr.contract

| Empty_implicit_contract of Tezos_protocol_environment_011_PtHangz2.Signature.Public_key_hash.t

| Empty_implicit_delegated_contract of Tezos_protocol_environment_011_PtHangz2.Signature.Public_key_hash.t

| Empty_transaction of Contract_repr.t

| Inconsistent_hash of Tezos_protocol_environment_011_PtHangz2.Signature.Public_key.t * Tezos_protocol_environment_011_PtHangz2.Signature.Public_key_hash.t * Tezos_protocol_environment_011_PtHangz2.Signature.Public_key_hash.t

| Inconsistent_public_key of Tezos_protocol_environment_011_PtHangz2.Signature.Public_key.t * Tezos_protocol_environment_011_PtHangz2.Signature.Public_key.t

| Failure of string

| Previously_revealed_key of Contract_repr.t

| Unrevealed_manager_key of Contract_repr.t

val exists : Raw_context.t -> Contract_repr.t -> bool Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

val must_exist : Raw_context.t -> Contract_repr.t -> unit Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

must_exist ctxt contract fails with the Non_existing_contract error if exists ctxt contract returns false. Even though this function is gas-free, it is always called in a context where some gas consumption is guaranteed whenever necessary. The first context is that of a transfer operation, and in that case the base cost of a manager operation (Micheclson_v1_gas.Cost_of.manager_operation) is consumed. The second context is that of an activation operation, and in that case no gas needs to be consumed since that operation is not a manager operation.

val allocated : Raw_context.t -> Contract_repr.t -> bool Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

val must_be_allocated : Raw_context.t -> Contract_repr.t -> unit Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

val list : Raw_context.t -> Contract_repr.t list Tezos_protocol_environment_011_PtHangz2.Lwt.t

val check_counter_increment : Raw_context.t -> Tezos_protocol_environment_011_PtHangz2.Signature.Public_key_hash.t -> Tezos_protocol_environment_011_PtHangz2.Z.t -> unit Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

val increment_counter : Raw_context.t -> Tezos_protocol_environment_011_PtHangz2.Signature.Public_key_hash.t -> Raw_context.t Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

val get_manager_key : Raw_context.t -> Tezos_protocol_environment_011_PtHangz2.Signature.Public_key_hash.t -> Tezos_protocol_environment_011_PtHangz2.Signature.Public_key.t Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

val is_manager_key_revealed : Raw_context.t -> Tezos_protocol_environment_011_PtHangz2.Signature.Public_key_hash.t -> bool Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

val reveal_manager_key : Raw_context.t -> Tezos_protocol_environment_011_PtHangz2.Signature.Public_key_hash.t -> Tezos_protocol_environment_011_PtHangz2.Signature.Public_key.t -> Raw_context.t Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

val get_balance : Raw_context.t -> Contract_repr.t -> Tez_repr.t Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

val get_balance_carbonated : Raw_context.t -> Contract_repr.t -> (Raw_context.t * Tez_repr.t) Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

val get_counter : Raw_context.t -> Tezos_protocol_environment_011_PtHangz2.Signature.Public_key_hash.t -> Tezos_protocol_environment_011_PtHangz2.Z.t Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

val get_script_code : Raw_context.t -> Contract_repr.t -> (Raw_context.t * Script_repr.lazy_expr option) Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

val get_script : Raw_context.t -> Contract_repr.t -> (Raw_context.t * Script_repr.t option) Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

val get_storage : Raw_context.t -> Contract_repr.t -> (Raw_context.t * Script_repr.expr option) Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

module Legacy_big_map_diff : sig ... end

val update_script_storage : Raw_context.t -> Contract_repr.t -> Script_repr.expr -> Lazy_storage_diff.diffs option -> Raw_context.t Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

val credit : Raw_context.t -> Contract_repr.t -> Tez_repr.t -> Raw_context.t Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

val spend : Raw_context.t -> Contract_repr.t -> Tez_repr.t -> Raw_context.t Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

val raw_originate : Raw_context.t -> ?prepaid_bootstrap_storage:bool -> Contract_repr.t -> balance:Tez_repr.t -> script:(Script_repr.t * Lazy_storage_diff.diffs option) -> delegate: Tezos_protocol_environment_011_PtHangz2.Signature.Public_key_hash.t option -> Raw_context.t Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

val fresh_contract_from_current_nonce : Raw_context.t -> (Raw_context.t * Contract_repr.t) Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult

val originated_from_current_nonce : since:Raw_context.t -> until:Raw_context.t -> Contract_repr.t list Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

val init : Raw_context.t -> Raw_context.t Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

val used_storage_space : Raw_context.t -> Contract_repr.t -> Tezos_protocol_environment_011_PtHangz2.Z.t Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

val paid_storage_space : Raw_context.t -> Contract_repr.t -> Tezos_protocol_environment_011_PtHangz2.Z.t Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

val set_paid_storage_space_and_return_fees_to_pay : Raw_context.t -> Contract_repr.t -> Tezos_protocol_environment_011_PtHangz2.Z.t -> (Tezos_protocol_environment_011_PtHangz2.Z.t * Raw_context.t) Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t