Module Tezos_raw_protocol_011_PtHangz2.Level_storage

val current : Raw_context.t -> Level_repr.t

val previous : Raw_context.t -> Level_repr.t

val root : Raw_context.t -> Level_repr.t

val from_raw : Raw_context.t -> Raw_level_repr.t -> Level_repr.t

val from_raw_with_offset : Raw_context.t -> offset:int32 -> Raw_level_repr.t -> Level_repr.t Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult

Fails with Negative_level_and_offset_sum if the sum of the raw_level and the offset is negative.

val pred : Raw_context.t -> Level_repr.t -> Level_repr.t option

val succ : Raw_context.t -> Level_repr.t -> Level_repr.t

val first_level_in_cycle : Raw_context.t -> Cycle_repr.t -> Level_repr.t

val last_level_in_cycle : Raw_context.t -> Cycle_repr.t -> Level_repr.t

val levels_in_cycle : Raw_context.t -> Cycle_repr.t -> Level_repr.t list

val levels_in_current_cycle : Raw_context.t -> ?offset:int32 -> unit -> Level_repr.t list

val levels_with_commitments_in_cycle : Raw_context.t -> Cycle_repr.t -> Level_repr.t list

val last_allowed_fork_level : Raw_context.t -> Raw_level_repr.t

val dawn_of_a_new_cycle : Raw_context.t -> Cycle_repr.t option

Returns Some cycle if the current level represents the last level of cycle and None if the level is not the last level of a cycle.

val may_snapshot_rolls : Raw_context.t -> bool

Returns true if the rolls should be snapshot at the current level.