Module Tezos_raw_protocol_011_PtHangz2.Main

Tezos Protocol Implementation - Protocol Signature Instance

This module is the entrypoint to the protocol for shells and other embedders. This signature is an instance of the Updater.PROTOCOL signature from the Protocol Environment.

Each Protocol depends on a version of the Protocol Environment. For the currently developed protocol, this is normally the latest version. You can see the full list of versions here.

For details on how Protocol and Environment interact, see this overview.

type validation_mode =

| Application of {

block_header : Alpha_context.Block_header.t;

baker : Alpha_context.public_key_hash;

}

| Partial_application of {

block_header : Alpha_context.Block_header.t;

baker : Alpha_context.public_key_hash;

}

| Partial_construction of {

predecessor : Tezos_protocol_environment_011_PtHangz2.Block_hash.t;

}

| Full_construction of {

predecessor : Tezos_protocol_environment_011_PtHangz2.Block_hash.t;

protocol_data : Alpha_context.Block_header.contents;

baker : Alpha_context.public_key_hash;

}

type validation_state = {

mode : validation_mode;

chain_id : Tezos_protocol_environment_011_PtHangz2.Chain_id.t;

ctxt : Alpha_context.t;

op_count : int;

migration_balance_updates : Alpha_context.Receipt.balance_updates;

liquidity_baking_escape_ema : Tezos_protocol_environment_011_PtHangz2.Int32.t;

implicit_operations_results : Apply_results.packed_successful_manager_operation_result list;

}

type operation_data = Alpha_context.packed_protocol_data

type operation = Alpha_context.packed_operation = {

shell : Tezos_protocol_environment_011_PtHangz2.Operation.shell_header;

protocol_data : operation_data;

}

val init_context : Tezos_protocol_environment_011_PtHangz2.Context.t -> Tezos_protocol_environment_011_PtHangz2.Context.t Tezos_protocol_environment_011_PtHangz2.Lwt.t

include Tezos_protocol_environment_011_PtHangz2.Updater.PROTOCOL with type block_header_data = Alpha_context.Block_header.protocol_data and type block_header_metadata = Apply_results.block_metadata and type block_header = Alpha_context.Block_header.t and type operation_data := operation_data and type operation_receipt = Apply_results.packed_operation_metadata and type operation := operation and type validation_state := validation_state

val max_block_length : int

val max_operation_data_length : int

val validation_passes : Tezos_protocol_environment_011_PtHangz2.Updater.quota list

type block_header_data = Alpha_context.Block_header.protocol_data

val block_header_data_encoding : block_header_data Tezos_protocol_environment_011_PtHangz2.Data_encoding.t

type block_header = Alpha_context.Block_header.t = {

shell : Tezos_protocol_environment_011_PtHangz2.Block_header.shell_header;

protocol_data : block_header_data;

}

type block_header_metadata = Apply_results.block_metadata

val block_header_metadata_encoding : block_header_metadata Tezos_protocol_environment_011_PtHangz2.Data_encoding.t

type operation_receipt = Apply_results.packed_operation_metadata

val operation_data_encoding : operation_data Tezos_protocol_environment_011_PtHangz2.Data_encoding.t

val operation_receipt_encoding : operation_receipt Tezos_protocol_environment_011_PtHangz2.Data_encoding.t

val operation_data_and_receipt_encoding : (operation_data * operation_receipt) Tezos_protocol_environment_011_PtHangz2.Data_encoding.t

val acceptable_passes : operation -> int list

val relative_position_within_block : operation -> operation -> int

val begin_partial_application : chain_id:Tezos_protocol_environment_011_PtHangz2.Chain_id.t -> ancestor_context:Tezos_protocol_environment_011_PtHangz2.Context.t -> predecessor_timestamp:Tezos_protocol_environment_011_PtHangz2.Time.t -> predecessor_fitness:Tezos_protocol_environment_011_PtHangz2.Fitness.t -> block_header -> validation_state Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

val begin_application : chain_id:Tezos_protocol_environment_011_PtHangz2.Chain_id.t -> predecessor_context:Tezos_protocol_environment_011_PtHangz2.Context.t -> predecessor_timestamp:Tezos_protocol_environment_011_PtHangz2.Time.t -> predecessor_fitness:Tezos_protocol_environment_011_PtHangz2.Fitness.t -> block_header -> validation_state Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

val begin_construction : chain_id:Tezos_protocol_environment_011_PtHangz2.Chain_id.t -> predecessor_context:Tezos_protocol_environment_011_PtHangz2.Context.t -> predecessor_timestamp:Tezos_protocol_environment_011_PtHangz2.Time.t -> predecessor_level:Tezos_protocol_environment_011_PtHangz2.Int32.t -> predecessor_fitness:Tezos_protocol_environment_011_PtHangz2.Fitness.t -> predecessor:Tezos_protocol_environment_011_PtHangz2.Block_hash.t -> timestamp:Tezos_protocol_environment_011_PtHangz2.Time.t -> ?protocol_data:block_header_data -> unit -> validation_state Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

val apply_operation : validation_state -> operation -> (validation_state * operation_receipt) Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

val finalize_block : validation_state -> Tezos_protocol_environment_011_PtHangz2.Block_header.shell_header option -> (Tezos_protocol_environment_011_PtHangz2.Updater.validation_result * block_header_metadata) Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

val rpc_services : Tezos_protocol_environment_011_PtHangz2.Updater.rpc_context Tezos_protocol_environment_011_PtHangz2.RPC_directory.t

val init : Tezos_protocol_environment_011_PtHangz2.Context.t -> Tezos_protocol_environment_011_PtHangz2.Block_header.shell_header -> Tezos_protocol_environment_011_PtHangz2.Updater.validation_result Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

val value_of_key : chain_id:Tezos_protocol_environment_011_PtHangz2.Chain_id.t -> predecessor_context:Tezos_protocol_environment_011_PtHangz2.Context.t -> predecessor_timestamp:Tezos_protocol_environment_011_PtHangz2.Time.t -> predecessor_level:Tezos_protocol_environment_011_PtHangz2.Int32.t -> predecessor_fitness:Tezos_protocol_environment_011_PtHangz2.Fitness.t -> predecessor:Tezos_protocol_environment_011_PtHangz2.Block_hash.t -> timestamp:Tezos_protocol_environment_011_PtHangz2.Time.t -> (Tezos_protocol_environment_011_PtHangz2.Context.Cache.key -> Tezos_protocol_environment_011_PtHangz2.Context.Cache.value Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t) Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t