Module Tezos_raw_protocol_011_PtHangz2.Services_registration

Functions for RPC service registration, using Updater.rpc_context and RPC_service.t from the Protocol Environment.

This module is a frontend to a mutable service directory. The various register functions update the directory as a side-effect.

The get_rpc_services function returns the resulting RPC_context. It is parameterized by Updater.rpc_context which acts as the service prefix (in practice meaning this type will be passed to each handler). Hence, Protocol RPC services provide a read-only view of the Ledger state.

type rpc_context = {

block_hash : Tezos_protocol_environment_011_PtHangz2.Block_hash.t;

block_header : Alpha_context.Block_header.shell_header;

context : Alpha_context.t;

}

val rpc_init : Tezos_protocol_environment_011_PtHangz2.Updater.rpc_context -> rpc_context Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t

val register0 : chunked:bool -> ([< Tezos_protocol_environment_011_PtHangz2.RPC_service.meth ], Tezos_protocol_environment_011_PtHangz2.Updater.rpc_context, Tezos_protocol_environment_011_PtHangz2.Updater.rpc_context, 'a, 'b, 'c) Tezos_protocol_environment_011_PtHangz2.RPC_service.t -> (Alpha_context.t -> 'a -> 'b -> 'c Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t) -> unit

val register0_noctxt : chunked:bool -> ([< Tezos_protocol_environment_011_PtHangz2.RPC_service.meth ], Tezos_protocol_environment_011_PtHangz2.Updater.rpc_context, 'a, 'b, 'c, 'd) Tezos_protocol_environment_011_PtHangz2.RPC_service.t -> ('b -> 'c -> 'd Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t) -> unit

val register1 : chunked:bool -> ([< Tezos_protocol_environment_011_PtHangz2.RPC_service.meth ], Tezos_protocol_environment_011_PtHangz2.Updater.rpc_context, Tezos_protocol_environment_011_PtHangz2.Updater.rpc_context * 'a, 'b, 'c, 'd) Tezos_protocol_environment_011_PtHangz2.RPC_service.t -> (Alpha_context.t -> 'a -> 'b -> 'c -> 'd Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t) -> unit

val register2 : chunked:bool -> ([< Tezos_protocol_environment_011_PtHangz2.RPC_service.meth ], Tezos_protocol_environment_011_PtHangz2.Updater.rpc_context, (Tezos_protocol_environment_011_PtHangz2.Updater.rpc_context * 'a) * 'b, 'c, 'd, 'e) Tezos_protocol_environment_011_PtHangz2.RPC_service.t -> (Alpha_context.t -> 'a -> 'b -> 'c -> 'd -> 'e Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t) -> unit

val opt_register0 : chunked:bool -> ([< Tezos_protocol_environment_011_PtHangz2.RPC_service.meth ], Tezos_protocol_environment_011_PtHangz2.Updater.rpc_context, Tezos_protocol_environment_011_PtHangz2.Updater.rpc_context, 'a, 'b, 'c) Tezos_protocol_environment_011_PtHangz2.RPC_service.t -> (Alpha_context.t -> 'a -> 'b -> 'c option Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t) -> unit

val opt_register1 : chunked:bool -> ([< Tezos_protocol_environment_011_PtHangz2.RPC_service.meth ], Tezos_protocol_environment_011_PtHangz2.Updater.rpc_context, Tezos_protocol_environment_011_PtHangz2.Updater.rpc_context * 'a, 'b, 'c, 'd) Tezos_protocol_environment_011_PtHangz2.RPC_service.t -> (Alpha_context.t -> 'a -> 'b -> 'c -> 'd option Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t) -> unit

val opt_register2 : chunked:bool -> ([< Tezos_protocol_environment_011_PtHangz2.RPC_service.meth ], Tezos_protocol_environment_011_PtHangz2.Updater.rpc_context, (Tezos_protocol_environment_011_PtHangz2.Updater.rpc_context * 'a) * 'b, 'c, 'd, 'e) Tezos_protocol_environment_011_PtHangz2.RPC_service.t -> (Alpha_context.t -> 'a -> 'b -> 'c -> 'd -> 'e option Tezos_protocol_environment_011_PtHangz2.Error_monad.tzresult Tezos_protocol_environment_011_PtHangz2.Lwt.t) -> unit

val get_rpc_services : unit -> Tezos_protocol_environment_011_PtHangz2.Updater.rpc_context Tezos_protocol_environment_011_PtHangz2.RPC_directory.directory