Module Contract.Inactive_delegate

type t = Raw_context.t

type context = t

type elt = Contract_repr.t

The type of elements.

val mem : context -> elt -> bool Tezos_protocol_environment_011_PtHangz2.Lwt.t

Tells if a elt is a member of the set

val add : context -> elt -> Raw_context.t Tezos_protocol_environment_011_PtHangz2.Lwt.t

Adds a elt is a member of the set

val remove : context -> elt -> Raw_context.t Tezos_protocol_environment_011_PtHangz2.Lwt.t

Removes a elt of the set ; does nothing if not a member

val elements : context -> elt list Tezos_protocol_environment_011_PtHangz2.Lwt.t

Returns the elements of the set, deserialized in a list in no particular order.

val fold : context -> init:'a -> f:(elt -> 'a -> 'a Tezos_protocol_environment_011_PtHangz2.Lwt.t) -> 'a Tezos_protocol_environment_011_PtHangz2.Lwt.t

Iterates over the elements of the set.

val clear : context -> Raw_context.t Tezos_protocol_environment_011_PtHangz2.Lwt.t

Removes all elements in the set