Module Tezos_raw_protocol_011_PtHangz2.Storage_sigs

Entity Accessor Signatures

module type Single_data_storage = sig ... end

The generic signature of a single data accessor (a single value bound to a specific key in the hierarchical (key x value) database).

module type Non_iterable_indexed_data_storage = sig ... end

Restricted version of Indexed_data_storage w/o iterators.

module type Non_iterable_indexed_carbonated_data_storage = sig ... end

Variant of Non_iterable_indexed_data_storage with gas accounting.

module type Non_iterable_indexed_carbonated_data_storage_with_values = sig ... end

module type Non_iterable_indexed_carbonated_data_storage_INTERNAL = sig ... end

module type Indexed_data_storage = sig ... end

The generic signature of indexed data accessors (a set of values of the same type indexed by keys of the same form in the hierarchical (key x value) database).

module type Indexed_data_snapshotable_storage = sig ... end

module type Data_set_storage = sig ... end

The generic signature of a data set accessor (a set of values bound to a specific key prefix in the hierarchical (key x value) database).

module type Carbonated_data_set_storage = sig ... end

Variant of Data_set_storage with gas accounting.

module type NAME = sig ... end

module type VALUE = sig ... end

module type REGISTER = sig ... end

module type Indexed_raw_context = sig ... end